==> notebooks/00_update_local_data.ipynb <==
This is a simple utility notebook to update the games.csv file with the latest data from NBA.com. You want to run this notebook before backfilling the data to Hopsworks.ai for the first time.

import os

import pandas as pd

from datetime import datetime, timedelta
from pytz import timezone

# change working directory to project root when running from notebooks folder to make it easier to import modules
# and to access
# sibling folders
os.chdir("..")

from src.utils.webscraping import (
    activate_web_driver,
    scrape_to_dataframe,
    convert_columns,
    combine_home_visitor,
)

from pathlib import Path

DATAPATH = Path(r"data")

pd.set_option("display.max_columns", None)

**Determine How current is the existing data**

games_old = pd.read_csv(DATAPATH / "raw" / "games.csv")

# Find the last date and season in the current dataset
last_date = games_old["GAME_DATE_EST"].max()
last_season = games_old["SEASON"].max()

# remove the time from the date
last_date = last_date.split(" ")[0]

# Determine the date of the next day to begin scraping from
start_date = (datetime.strptime(last_date, "%Y-%m-%d") + timedelta(days=1)).date()

# determine what season we are in currently
today = datetime.now(timezone("EST"))  # nba.com uses US Eastern Standard Time
if today.month >= 10:
    current_season = today.year
else:
    current_season = today.year - 1

# determine which seasons we need to scrape to catch up the data
seasons = list(range(last_season, current_season + 1))


print("Last date in dataset: ", last_date)
print("Last season in dataset: ", last_season)
print("Current season: ", current_season)
print("Seasons to scrape: ", seasons)
print("Start date: ", start_date)

# if the last date in the dataset is today, then we don't need to scrape any new data
if start_date > datetime.now().date():
    print("No new data to scrape")
    exit()

**Activate Webdriver**

# initiate a webdriver in selenium
# since website data is dynamically generated

driver = activate_web_driver()

**Scrape New Completed Games and Format Them**

season_types = ["Regular+Season", "PlayIn", "Playoffs"]
all_season_types = pd.DataFrame()

def update_games(driver, season, start_date, end_date) -> pd.DataFrame:

    season_types = ["Regular+Season", "PlayIn", "Playoffs"]

    all_season_types = pd.DataFrame()

    for season_type in season_types:

        df = scrape_to_dataframe(
            api_key="",
            driver=driver,
            Season=season,
            DateFrom=start_date,
            DateTo=end_date,
            season_type=season_type,
        )

        if not (df.empty):
            df = convert_columns(df)
            df = combine_home_visitor(df)
            all_season_types = pd.concat([all_season_types, df], axis=0)

    return all_season_types

new_games = pd.DataFrame()

for season in seasons:
    end_date = datetime.strptime(
        f"{season+1}-08-01", "%Y-%m-%d"
    ).date()  # use August 1st to get all games from the current season
    print(f"Scraping season {season} from {start_date} to {end_date}")
    df_season = update_games(driver, str(season), str(start_date), str(end_date))
    new_games = pd.concat([new_games, df_season], axis=0)
    start_date = datetime.strptime(
        f"{season+1}-10-01", "%Y-%m-%d"
    ).date()  # if more than 1 season, reset start date to beginning of next season


new_games

**Close Webdriver**

driver.close()

**Append to Games.csv**

new_games["GAME_STATUS_TEXT"] = "Final"

games_old = games_old.drop(
    columns=["TEAM_ID_home", "TEAM_ID_away"], axis=1
)

games = pd.concat([games_old, new_games], axis=0)

games.to_csv(DATAPATH / "interim" / "games.csv", index=False)

games
