==> notebooks/02_data_processing.ipynb <==
# Data Processing

import pandas as pd
import numpy as np

pd.set_option('display.max_columns', 500)

# For Visualization
import matplotlib.pyplot as plt
import seaborn as sns

from pathlib import Path  #for Windows/Linux compatibility
DATAPATH = Path(r'../data')


## games.csv

games = pd.read_csv(DATAPATH / "interim" /"games.csv")
games.head()

games.shape

**Clean Data**

#remove preseason games (GAME_ID begins with a 1)
games = games[games['GAME_ID'] > 20000000]

#flag postseason games (GAME_ID begins with >2)
games['PLAYOFF'] = (games['GAME_ID'] >= 30000000).astype('int8')

#remove duplicates (each GAME_ID should be unique)
games = games[~games.duplicated(subset=['GAME_ID'])]

#drop unnecessary fields
drop_fields = ['GAME_STATUS_TEXT']
games = games.drop(drop_fields,axis=1)

games
    

games["TARGET"] = games["HOME_TEAM_WINS"]

games

**Save Train Data**

games.to_csv(DATAPATH / "interim" / "transformed.csv", index=False)
