==> notebooks/03_train_test_split.ipynb <==
### Train / Test Split

This notebook splits the data into train data and test data.

It selects the latest season and uses this as the test data.

import pandas as pd
import numpy as np

from pathlib import Path  # for Windows/Linux compatibility

DATAPATH = Path(r"../data")

df = pd.read_csv(DATAPATH / "interim" / "transformed.csv")

**Split Data to Train and Test**

latest_season = df["SEASON"].unique().max()

train = df[df["SEASON"] < (latest_season)]
test = df[df["SEASON"] >= (latest_season - 1)]

train.to_csv(DATAPATH / "processed" / "train.csv", index=False)
test.to_csv(DATAPATH / "processed" / "test.csv", index=False)
